Move `willModifyDb` and `generateSuccessMessage` into `dolt_checkout_helprs.go`
