+++ go/libraries/doltcore/sqle/dprocedures/dolt_checkout_helpers.go
@@ -17,6 +17,7 @@ package dprocedures
 import (
 	"errors"
 	"fmt"
+	"github.com/dolthub/dolt/go/libraries/doltcore/env"
 
 	"github.com/dolthub/go-mysql-server/sql"
 
@@ -188,4 +189,40 @@ func transferWorkingChanges(
 	}
 
 	return nil
+}
+
+// willModifyDb determines whether or not this operation is a no-op and can return early with a helpful message.
+func willModifyDb(dSess *dsess.DoltSession, data env.DbData, dbName, branchName string, updateHead bool) (bool, error) {
+	headRef, err := data.Rsr.CWBHeadRef()
+	// If we're in a detached head state, allow checking out a new branch.
+	if err == doltdb.ErrOperationNotSupportedInDetachedHead {
+		return true, nil
+	}
+	if err != nil {
+		return false, err
+	}
+
+	// If the operation won't modify either the active session or the default session, return early.
+	isModification := headRef.GetPath() != branchName
+	if updateHead {
+		fs, err := dSess.Provider().FileSystemForDatabase(dbName)
+		if err != nil {
+			return false, err
+		}
+		repoState, err := env.LoadRepoState(fs)
+		if err != nil {
+			return false, err
+		}
+		defaultBranch := repoState.CWBHeadRef().GetPath()
+		isModification = isModification || (defaultBranch != branchName)
+	}
+	return isModification, nil
+}
+
+func generateSuccessMessage(newBranch, upstream string) string {
+	result := fmt.Sprintf("Switched to branch '%s'", newBranch)
+	if upstream != "" {
+		result += fmt.Sprintf("\nbranch '%s' set up to track '%s'.", newBranch, upstream)
+	}
+	return result
 }
